Fix patch format for debian/patches/0006-Output-addressType-public-vs.-random-when-discoverin.patch
The previous patch was generated in a Candor environment, which has a different quilt version than on Veracity.
As a result, the patch format generated by `gbq-pq export` was different, but the patch still applied.
Re-export the patch to get it into a format matching all of the other patches.